services/api-gateway/requirements.txt
typing-extensions==4.15.0
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.5
flask==3.1.2
gunicorn
psutil==7.2.2
redis==7.3.0